4.16.0.71:
Revision 1.421  2010/05/17 02:38:38  hugh
- add new variable display-default-filenames (Casey Leedom)
- eliminate most GCC warnings; improve handling of some errors
- allow for Linux/glibc elimination of I_PUSH (pseudo TTY STREAMS)
- improve jove.spec for Red Hat packaging
- delete obsolete command process-dbx-output
- delete obsolete variables allow-bad-filenames, display-bad-filenames, internal-tabstop
- add bindings for more xterm function key variants

+++ fp.c
@@ -94,7 +94,7 @@ int	flags,
 	case F_APPEND:
 		fd = open(name, O_WRONLY | O_BINARY);
 		if (fd != -1) {
-			(void) lseek(fd, 0L, 2);
+			(void) lseek(fd, (off_t)0, 2);
 			break;
 		}
 		/* FALLTHROUGH */
@@ -222,7 +222,7 @@ off_t	offset;
 		flushout(fp);
 	fp->f_cnt = 0;		/* next read will f_filbuf(), next write
 				   will flush() with no bad effects */
-	lseek(fp->f_fd, (long) offset, L_SET);
+	lseek(fp->f_fd, offset, L_SET);
 }
 
 void
